+++ drift_flutter/lib/drift_flutter.dart
@@ -15,14 +15,16 @@ import 'src/connect.dart' as connect;
 /// }
 /// ```
 ///
-/// [name] is the name of the database to use. On native platforms, this is the
-/// file name of the database without an extension.
+/// [name] is the name of the database to use. On native platforms, a file
+/// called `$name.sqlite` in `getApplicationDocumentsDirectory()` will be used
+/// for the database.
 /// On the web, this name is part of the FileSystem API path or the name of the
 /// IndexedDB database to use.
 /// Typically, names only consist of alphanumerical characters and underscores.
 QueryExecutor driftDatabase({
   required String name,
   connect.DriftWebOptions? web,
+  connect.DriftNativeOptions? native,
 }) {
-  return connect.driftDatabase(name: name, web: web);
+  return connect.driftDatabase(name: name, web: web, native: native);
 }