+++ drift_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+- Add `DriftNativeOptions` with `shareAcrossIsolates` option that will give
+  multiple isolates access to the same drift database without having to manually
+  set up ports.
+
 ## 0.1.0
 
 - Initial version.